update image logic
