+++ src/main/java/by/dudko/carsales/model/entity/BaseEntity.java
@@ -2,5 +2,6 @@ package by.dudko.carsales.model.entity;
 
 public interface BaseEntity<T> {
     T getId();
+
     void setId(T id);
 }
